+++ config/settings_ci.py
@@ -64,6 +64,7 @@ DATABASES = {
         "HOST": "postgres",
         "PORT": "5432",
         "ENGINE": "psqlextra.backend",
+        "ATOMIC_REQUESTS": True,  # perform HTTP requests as atomic transactions
         "OPTIONS": {
             # prevent libpq from automatically trying to connect to the db via GSSAPI
             "gssencmode": "disable",